# cmd/main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для библиотеки GOtus.",
        "title": "GOtus API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/booking/cancel/{id}": {
            "post": {
                "description": "Отменяет бронирование по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Отменить бронирование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование отменено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/booking/create": {
            "post": {
                "description": "Бронирует доступный экземпляр книги по ISBN для указанного пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Создать бронирование книги",
                "parameters": [
                    {
                        "description": "Данные для бронирования",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_booking.CreateBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Бронирование успешно создано",
                        "schema": {
                            "$ref": "#/definitions/internal_api_booking.CreateBookingResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/booking/end/{id}": {
            "post": {
                "description": "Завершает бронирование по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Завершить бронирование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование завершено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/booking/extend/{id}": {
            "put": {
                "description": "Продлевает бронирование по ID на заданное количество дней",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "booking"
                ],
                "summary": "Продлить бронирование",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID бронирования",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Дни продления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_booking.ExtendBookingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование продлено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api_booking.CreateBookingRequest": {
            "type": "object",
            "properties": {
                "isbn": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_api_booking.CreateBookingResponse": {
            "type": "object",
            "properties": {
                "booking_id": {
                    "type": "integer"
                }
            }
        },
        "internal_api_booking.ExtendBookingRequest": {
            "type": "object",
            "properties": {
                "extension_days": {
                    "type": "integer"
                }
            }
        },
...
    }
}